split debug/check endpoints consolidate health-servlet
