--- metadata-service/health-servlet/build.gradle
@@ -1,22 +1,0 @@
-apply plugin: 'java'
-
-dependencies {
-
-  implementation project(':metadata-service:factories')
-
-  implementation externalDependency.guava
-  implementation externalDependency.reflections
-  implementation externalDependency.springBoot
-  implementation externalDependency.springCore
-  implementation externalDependency.springDocUI
-  implementation externalDependency.springWeb
-  implementation externalDependency.springWebMVC
-  implementation externalDependency.springBeans
-  implementation externalDependency.springContext
-  implementation externalDependency.slf4jApi
-  compileOnly externalDependency.lombok
-  implementation externalDependency.antlr4Runtime
-  implementation externalDependency.antlr4
-
-  annotationProcessor externalDependency.lombok
-}
